+++ src/lib/modules/SATModule/SATModule.h
@@ -532,11 +532,11 @@ namespace smtrat
 
     inline void SATModule::claBumpActivity( Minisat::Clause& c )
     {
-        if( (c.activity() += cla_inc) > 1e20 )
+        if( (c.activity() += (float)cla_inc) > 1e20 )
         {
             // Rescale:
             for( int i = 0; i < learnts.size(); i++ )
-                ca[learnts[i]].activity() *= 1e-20;
+                ca[learnts[i]].activity() *= (float)1e-20;
             cla_inc *= 1e-20;
         }
     }
